Redo gear score commit due to bad merge on logger
