+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "loa-details-v2",
-  "version": "0.3.9",
+  "version": "0.3.8",
   "description": "Damage meter for Lost Ark",
   "productName": "LOA Details",
   "author": "Eren Kara <erenkyi0@gmail.com>",
